storage/: Set file size limit instead of using mount
